[dtensor][1/n] refactor op dispatch logic to reduce overhead
This PR is the first change of a series of refactors to the op dispatch logic to:
1. remove the redundant logic in the op dispatch, simplify the error
checking
2. reduce the number of tree_map/tree_flatten/unflatten needed to reduce
the overhead coming from those operations
3. remove the CachedShardingPropagator by using lru_cache from functools
directly, this makes it not only helps TP, but general DTensor
operations could be faster!
4. change the view ops behavior by inplace changing the op_schema, which
is dangerous for sharding prop caching, model the view op as one type
of resharding too
5. enrich output sharding to include whether the op needs redistribute
so that we don't need explicit op schema comparison to know it.

This should help with further reducing the CPU overhead, benchmark
results coming soon

By refactoring the dispatch logic, it could possibly enable us to have
more features later (i.e. add IrregularShard placement easier)

ghstack-source-id: 37b8a2f2eba0a4a587fdfe1927cd9bf1188ada1d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107305